修复日志记录：修正size_t格式说明符并移除自动人脸识别
- 在TriggerFaceRecognition()内存检查日志中将格式说明符从%zu更正为%u，并使用显式类型转换
- 在CaptureJpegChunks()分配失败日志中将格式说明符从%zu更正为%u，并使用显式类型转换
- 在JPEG编码完成日志中将格式说明符从%zu更正为%u，并使用显式类型转换
- 移除门铃按下和PIR检测事件上的自动人脸识别触发
- 更新事件处理程序以记录事件而不触发识别，委托给服务器MCP调用
- 对assets.cc中的包含语句进行整理，以确保一致性和可读性
- 通过使用显式类型转换的unsigned int格式而不是C99的size_t格式，确保跨平台兼容性
- 在CHANGELOG.md中更新，详细记录日志格式修复及其技术理由

+++ main/protocols/protocol.cc
@@ -117,7 +117,7 @@ bool Protocol::IsTimeout() const {
       now - last_incoming_time_);
   bool timeout = duration.count() > kTimeoutSeconds;
   if (timeout) {
-    ESP_LOGE(TAG, "Channel timeout %ld seconds", (long)duration.count());
+    ESP_LOGE(TAG, "Channel timeout %d seconds", (int)duration.count());
   }
   return timeout;
 }
